same edit this page

+++ app/Http/Controllers/Admin/Product/CategoryController.php
@@ -62,42 +62,59 @@ class CategoryController extends Controller
 
     }
 
-    public function categoryUpdate(Request $request , $id)  {
-
+    public function categoryUpdate(Request $request, $id)
+    {
         try {
             $request->validate([
                 'name' => 'required',
                 'img_url' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:6144',
             ]);
+
             $category = Category::findOrFail($id);
+
             $category->name = $request->name;
 
             if ($request->hasFile('img_url')) {
+                if ($category->img_url && file_exists(public_path($category->img_url))) {
+                    unlink(public_path($category->img_url));
+                }
+
                 $image = $request->file('img_url');
                 $imageName = time() . '.' . $image->getClientOriginalExtension();
+                $image->move(public_path('uploads/categories/'), $imageName);
 
-                $image->move(public_path('uploads/category/'), $imageName);
-
-                $category->update([
-                    'img_url' => 'uploads/category/' . $imageName,
-                ]);
+                $category->img_url = 'uploads/categories/' . $imageName;
             }
 
             $category->save();
-            return redirect()->route('category-list')->with('success', 'Product updated successfully.');
 
+            return redirect()->route('category-list')->with('success', 'Category updated successfully.');
         } catch (\Exception $e) {
-            return back()->with('error', 'Something went wrong');
+
+            return back()->with('error', 'Something went wrong: ' . $e->getMessage());
         }
-
-
     }
+
+
+
+
     public function categoryDelete($id)
     {
         $category = Category::findOrFail($id);
+        if ($category->subCategories()->exists()) {
+
+            return redirect()->route('category-list')->with('error', 'Category Has Related Subcategories Not Deleted.');
+        }
+
+        if (file_exists(public_path($category->img_url))) {
+            unlink(public_path($category->img_url));
+        }
+
         $category->delete();
+
         return redirect()->route('category-list')->with('success', 'Category deleted successfully.');
     }
+
 
 
 }
